# apidocs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API for the CyVerse Discovery Environment: app discovery and launching, analysis management, and iRODS data access.\n\nTo authenticate: open the Authorize dialog and enter your username and password under BasicAuth,\nrun POST /login, then copy the access_token from the response into the BearerAuth value\n(prefixed with \"Bearer \") to call the other endpoints.",
        "title": "formation",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/apps": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Lists apps with optional filters. Date filters use the form \u003coperator\u003e\u003cISO-8601 date\u003e,\ne.g. '\u003e2025-09-29' or '\u003c=2024-12-31T23:59:59' (see docs/DATE_FILTERING.md).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Apps"
                ],
                "summary": "List apps available to the user",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 100,
                        "description": "Page size (1-1000)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Page offset",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Search term forwarded to the apps service",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "vice",
                            "interactive",
                            "de",
                            "osg",
                            "tapis"
                        ],
                        "type": "string",
                        "description": "Job type filter",
                        "name": "job_type",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Case-insensitive substring match on the description",
                        "name": "description",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Case-insensitive substring match on the integrator username",
                        "name": "integrator",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Date filter on the integration date",
                        "name": "integration_date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Date filter on the last-edited date",
                        "name": "edited_date",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "total and apps",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Invalid pagination value or date filter",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/apps/job-types": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Apps"
                ],
                "summary": "List supported job types",
                "responses": {
                    "200": {
                        "description": "job_types entries with name, description, and internal_name",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/apps/{system_id}/{app_id}/parameters": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Apps"
                ],
                "summary": "Get an app's parameter groups",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Execution system id (e.g. de)",
                        "name": "system_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "App UUID",
                        "name": "app_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "groups and overall_job_type",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Invalid app ID format",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "502": {
                        "description": "Apps service error (including unknown apps)",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
...
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        },
        "BearerAuth": {
            "description": "A Keycloak JWT prefixed with \"Bearer \". Obtain one via POST /login.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}